Change matrix element type from long double to double

+++ Matrix/include/task.hpp
@@ -4,6 +4,6 @@
 
 namespace HWMatrix
 {
-Matrix<long double> read_matrix(std::istream& input);
+Matrix<double> read_matrix(std::istream& input);
 void task(std::istream& input, std::ostream& output);
 }
